🚚 Rename: 공통 레이아웃을 app/ui로 이동
